Move interfaces/events to modules/events (Fix #105)
As events.ts ended up containing some JS code after
transpilation for its enum this file needs to be a and
in modules.
