+++ src/interfaces/index.ts
@@ -5,6 +5,5 @@ export { IClientOptions } from './client-options';
 export { ICache } from './cache';
 export { IExchange, IExchangeResult } from './exchange';
 export { IGraphQLError, IGraphQLErrorInput } from './error';
-export { ClientEventType, ClientEvent } from './events';
 export { IOperation } from './operation';
 export { ISubscriptionObserver, ISubscription } from './subscription';
