+++ src/modules/events.ts
@@ -1,17 +1,17 @@
-import { IExchangeResult } from './exchange';
+import { IExchangeResult } from '../interfaces';
 
 export enum ClientEventType {
   InvalidateTypenames = 'InvalidateTypenames',
   RefreshAll = 'RefreshAll',
 }
 
-export interface IClientEvent<E, P> {
+export interface IClientEventRaw<E, P> {
   payload: P;
   type: E;
 }
 
-export type ClientEvent = IClientEvent<
+export type IClientEvent = IClientEventRaw<
   ClientEventType.InvalidateTypenames,
   { typenames: string[]; changes: IExchangeResult }
 > &
-  IClientEvent<ClientEventType.RefreshAll, void>;
+  IClientEventRaw<ClientEventType.RefreshAll, void>;
